requirements.txt
torch~=2.1.0+cu118
numpy~=1.24.1
tqdm~=4.64.1
pillow~=9.3.0
yaml~=0.2.5
pyyaml~=6.0.1
torchvision~=0.16.0+cu118
safetensors~=0.4.0
matplotlib~=3.5.3
loralib~=0.1.1